requirements.txt
Django 
djangorestframework 
django-cors-headers 
django-crispy-forms 
crispy-bootstrap5 
django-filter 
djangorestframework_simplejwt 
drf-spectacular 

# PostgreSQL Database Support
psycopg2-binary 

# Environment Variables Management
django-environ 

# CAPTCHA
django-captcha 
django-recaptcha 

# Whitenoise for static file serving
whitenoise 

requests
pillow 
pillow_heif 
pipenv 
drf-spectacular